restructure core + imageItem data
