+++ src/wezel/__init__.py
@@ -1,7 +1,6 @@
-from wezel.core import *
-from wezel.main import *
+from wezel.core import Action
+from wezel.main import app
 import wezel.actions
-import wezel.apps
 import wezel.widgets
 import wezel.menus
