--- src/wezel/apps/__init__.py
@@ -1,3 +1,0 @@
-
-from .dicom import *
-from .welcome import *
