+++ src/wezel/core.py
@@ -1,12 +1,19 @@
-__all__ = ['Main', 'App', 'Action', 'MenuBar', 'Menu']
+
 
 import sys
+import logging
 
-from PyQt5.QtCore import *
-from PyQt5.QtWidgets import QMainWindow, QAction, QMenu, QMenuBar
+#from PyQt5.QtCore import *
+from PyQt5.QtCore import Qt
+from PyQt5.QtWidgets import QApplication, QMainWindow, QAction, QMenu, QMenuBar, QDockWidget
 from PyQt5.QtGui import QIcon
 
-import wezel.icons as icons
+import dbdicom as db
+import wezel
+
+QApplication.setAttribute(Qt.AA_EnableHighDpiScaling, True)
+QApplication.setAttribute(Qt.AA_UseHighDpiPixmaps, True)
+
 
 # Examples of style sheets
 # https://doc.qt.io/qtforpython/overviews/stylesheet-examples.html
@@ -50,55 +57,16 @@ STYLESHEET = """
     """
 
 
-class Main(QMainWindow):
+class Wezel:
 
-    def __init__(self, wezel): 
-        """Creates the main window."""
+    def __init__(self):
+        self.log = logger()
+        self.QApp = QApplication(sys.argv)
+        self.QApp.setWindowIcon(QIcon(wezel.icons.favicon))
+        self.main = Main(self)
 
-        super().__init__()
-
-        self.wezel = wezel
-        #self.setStyleSheet(STYLESHEET)
-        self.setWindowTitle("Wezel")
-        self.setWindowIcon(QIcon(icons.favicon))
-
-    def closeEvent(self, event): 
-        accept = self.wezel.app.close()
-        if accept:
-            event.accept()
-        else:
-            event.ignore()
-
-
-class App:
-    """Base class for all Wezel applications"""
-
-    def __init__(self, wezel):
-        self.wezel = wezel
-
-    @property
-    def QApp(self):
-        return self.wezel.QApp
-
-    @property
-    def main(self):
-        return self.wezel.main
-
-    @property
-    def log(self):
-        return self.wezel.log
-
-    @property
-    def dialog(self):
-        return self.wezel.dialog
-
-    @property
-    def menubar(self):
-        return self.main.menuBar()
-
-    @property
-    def status(self):
-        return self.main.statusBar()
+    def open(self, path):
+        self.main.open(path)
 
     def show(self):    
         self.log.info('Launching Wezel!')
@@ -107,38 +75,192 @@ class App:
             self.QApp.exec()
             #sys.exit(self.QApp.exec())
         except Exception as e:
+            print('Error: ' + str(e))
             self.log.exception('Error: ' + str(e))
 
+
+class Main(QMainWindow):
+
+    def __init__(self, wzl): 
+        super().__init__()
+        self.wezel = wzl
+        #self.setStyleSheet(STYLESHEET)
+        self.setWindowTitle("Wezel")
+        #self.setWindowIcon(QIcon(wezel.icons.favicon))
+
+        self.dialog = wezel.widgets.Dialog(self)
+        self.status = wezel.widgets.StatusBar()
+        self.setStatusBar(self.status)
+
+        self.toolBar = wezel.canvas.ToolBar()
+        self.toolBarDockWidget = QDockWidget()
+        self.toolBarDockWidget.setWidget(self.toolBar)
+        self.addDockWidget(Qt.RightDockWidgetArea, self.toolBarDockWidget)
+        self.toolBarDockWidget.hide()
+
+        self.treeView = None
+        self.treeViewDockWidget = QDockWidget()
+        self.addDockWidget(Qt.LeftDockWidgetArea, self.treeViewDockWidget)
+        self.treeViewDockWidget.hide()
+
+        self.folder = None
+        #self.central = QSplitter()
+        self.central = wezel.widgets.MainMultipleDocumentInterface()
+        self.central.subWindowActivated.connect(lambda subWindow: self.setSubWindow(subWindow))
+        self.setCentralWidget(self.central)
+
+        self.set_menu(wezel.menus.dicom)
+
+    def closeEvent(self, event): #main
+        accept = self.close()
+        if accept:
+            event.accept()
+        else:
+            event.ignore()
+
     def set_menu(self, menu):
-        menubar = MenuBar(self, menu)
-        self.main.setMenuBar(menubar)
+        self.menubar = MenuBar(self, menu)
+        self.setMenuBar(self.menubar)
 
-    def set_central(self, widget):
-        self.main.setCentralWidget(widget)
+    def setSubWindow(self, subWindow):
+        # activeWindow = self.central.activeWindow
+        # if activeWindow is not None:
+        #     activeWidget = activeWindow.widget()
+        #     if activeWidget.__class__.__name__ == 'SeriesCanvas':
+        #         activeWidget.regions = self.toolBar.regionList.regions
+        self.central.activeWindow = subWindow
+        if self.folder is None:
+            return
+        widget = subWindow.widget()
+        if widget.__class__.__name__ == 'SeriesCanvas':
+            self.toolBar.setSeriesCanvas(widget)
 
-    def set_status(self, message):
-        self.status.message(message)
-
-    def set_data(self, data=None):
-        pass
-
-    def get_data(self):
-        pass
-
-    def set_app(self, App):
-        wezel = self.wezel
-        wezel.app = App(wezel) 
-        self.__class__ = App 
-        self.__dict__ = wezel.app.__dict__
-        return wezel.app   
+    def open(self, path):
+        self.folder = db.database(path=path, 
+            status = self.status, 
+            dialog = self.dialog)
+        self.display(self.folder)
 
     def close(self):
-        """
-        Called when the user attemps to close the main window.
+        """Closes the application."""
+        if self.folder is None:
+            return True
+        accept = self.folder.close()
+        if accept:
+            self.folder = None
+            self.toolBarDockWidget.hide()
+            self.treeViewDockWidget.hide()
+            for subWindow in self.central.subWindowList():
+                self.central.removeSubWindow(subWindow)
+            #self.central.closeAllSubWindows()
+            self.menuBar().enable()
+        #    self.set_app(apps.WezelWelcome)
+        return accept
 
-        Returns: True if the window can be closed, False if not.
+    def refresh(self):
         """
-        return True
+        Refreshes the Wezel display.
+        """
+        self.status.message('Refreshing display..')
+        self.treeView.setFolder()
+        self.menuBar().enable()
+        self.status.hide()
+        #self.status.message()
+
+    def addAsSubWindow(self, widget, title=None, icon=None):
+        """
+        displays a widget as a subwindow in the MDI. 
+        
+        Returns the subwindow
+        """ 
+        self.central.addWidget(widget, title=title, icon=icon)
+        
+    def display(self, object):
+
+        if object.type() == 'Database':
+            self.treeView = wezel.widgets.DICOMFolderTree(object)
+            self.treeView.itemSelectionChanged.connect(self.menuBar().enable)
+            self.treeViewDockWidget.setWidget(self.treeView)
+            self.treeViewDockWidget.show()
+            self.menuBar().enable()
+        elif object.type() == 'Patient': # No Patient Viewer yet
+            pass
+        elif object.type() == 'Study': # No Study Viewer yet
+            pass
+        elif object.type() == 'Series':
+            seriesCanvas = wezel.widgets.SeriesCanvas()
+            seriesCanvas.setImageSeries(object)
+            seriesCanvas.mousePositionMoved.connect(
+                lambda x, y: self.status.pixelValue(x,y,seriesCanvas.array())
+            )
+            self.addAsSubWindow(seriesCanvas, title=object.label())
+            self.central.tileSubWindows()
+            self.toolBar.setSeriesCanvas(seriesCanvas)
+            self.toolBarDockWidget.show()
+        elif object.type() == 'Instance':
+            pass
+
+    def get_selected(self, generation):   
+        if self.treeView is None: 
+            return []
+        return self.treeView.get_selected(generation)
+
+    def selected(self, generation):
+        if self.treeView is None: 
+            return []
+        return self.treeView.selected(generation)
+ 
+    def nr_selected(self, generation):
+        if self.treeView is None: 
+            return 0
+        return self.treeView.nr_selected(generation)
+    
+    
+    # def set_data(self, folder):
+    #     if self.folder is not None:
+    #         if self.folder.close():
+    #             self.central.closeAllSubWindows()
+    #         else:
+    #             return
+    #     self.folder = folder
+    #     if self.treeView is None:
+    #         self.display(folder)
+    #     else:
+    #         self.treeView.setFolder(folder)
+    #     self.menubar.enable()
+
+    # def addSubWindow(self, subWindow):
+
+    #     self.central.addSubWindow(subWindow) 
+
+    # def closeAllSubWindows(self):
+    #     """
+    #     Closes all open windows.
+    #     """
+    #     self.central.closeAllSubWindows()
+
+    # def closeSubWindow(self, subWindowName):
+    #     """
+    #     Closes all subwindows with a given name
+    #     """   
+    #     self.central.closeSubWindow(subWindowName)
+
+    # def tileSubWindows(self):
+    #     """
+    #     Tiles all open windows.
+    #     """
+    #     self.central.tileSubWindows()
+
+    # def addAsDockWidget(self, widget):
+
+    #     #dockwidget = QDockWidget(self.main, Qt.SubWindow)
+    #     dockwidget = QDockWidget()
+    #     #dockwidget.setAllowedAreas(Qt.LeftDockWidgetArea)
+    #     #dockwidget.setWindowTitle('DICOM database')
+    #     #dockwidget.setFeatures(QDockWidget.NoDockWidgetFeatures)
+    #     #dockwidget.setObjectName(widget.__class__.__name__)
+    #     dockwidget.setWidget(widget)
+    #     self.main.addDockWidget(Qt.LeftDockWidgetArea, dockwidget)
 
 
 class MenuBar(QMenuBar):
@@ -146,11 +268,11 @@ class MenuBar(QMenuBar):
     Programming interfaces for the Wezel menus. 
     """
 
-    def __init__(self, app, menu):
+    def __init__(self, main, menu):
         super().__init__()
 
         self._menus = []
-        self.app = app
+        self.main = main
         menu(self)
         self.enable()
 
@@ -174,13 +296,13 @@ class MenuBar(QMenuBar):
 
 class Menu(QMenu):
 
-    def __init__(self, parent=None, title='Menu'):
+    def __init__(self, parent, title='Menu'):
         super().__init__()
 
         self._actions = []
         self._menus = []
         self.setTitle(title)
-        self.app = parent.app
+        self.main = parent.main
         if parent is not None:
             parent.addMenu(self)
 
@@ -192,7 +314,11 @@ class Menu(QMenu):
         return Menu(self, title)
 
     def action(self, action, **kwargs):
-        return action(self, **kwargs)
+        #return action(self, **kwargs)
+        action = action(self, **kwargs)
+        self.addAction(action)
+        self._actions.append(action)
+        return action
         
     def separator(self):
         self.addSeparator() 
@@ -204,7 +330,7 @@ class Menu(QMenu):
         for submenu in self._menus:
             submenu.enable()
         for action in self._actions:
-            enable = action.enable(action.app)
+            enable = action.enable(action.main)
             action.setEnabled(enable)
 
 
@@ -220,20 +346,27 @@ class Action(QAction):
         """parent: App, Menu or MenuBar"""
         super().__init__()
 
-        if hasattr(parent, 'app'):
-            self.app = parent.app
-            parent.addAction(self)
-            parent._actions.append(self)
-        else:
-            self.app = parent
+        self.main = parent.main
+        # parent.addAction(self)
+        # parent._actions.append(self)
+
+        # if hasattr(parent, 'main'):
+        #     self.main = parent.main
+        #     parent.addAction(self)
+        #     parent._actions.append(self)
+        # else:
+        #     self.main = parent
         if text is None:
             text = self.__class__.__name__
         self.setText(text)
-        self.triggered.connect(lambda: self.run(self.app))
+        self.triggered.connect(lambda: self.run(self.main))
     
-        if icon is not None: self.setIcon(QIcon(icon))
-        if shortcut is not None: self.setShortcut(shortcut)
-        if tooltip is not None: self.setToolTip(tooltip)
+        if icon is not None: 
+            self.setIcon(QIcon(icon))
+        if shortcut is not None: 
+            self.setShortcut(shortcut)
+        if tooltip is not None: 
+            self.setToolTip(tooltip)
 
         # Dictionary with optional settings
         for option in kwargs:
@@ -244,4 +377,23 @@ class Action(QAction):
 
     def run(self, app):
         pass
+
+
+
+
+
+
+def logger():
+    
+    LOG_FILE_NAME = "wezel_log.log"
+    # creates some sort of conflict with mdreg - commenting out for now
+#    if os.path.exists(LOG_FILE_NAME):
+#        os.remove(LOG_FILE_NAME)
+    LOG_FORMAT = "%(levelname)s %(asctime)s - %(message)s"
+    logging.basicConfig(
+        filename = LOG_FILE_NAME, 
+        level = logging.INFO, 
+        format = LOG_FORMAT)
+    return logging.getLogger(__name__)
+
